Use better getter/setter parameter names
This is the result of a discussion with @mcflugen. We want to
convey meaning in the arguments for the getter and setter methods.
Using "dest_ptr" helps clarify that a reference to a data array is
being returned through this *get_value_ptr* parameter. The "src"
parameter name in *set_value* is congruent with *set_value_at_indices*
and incongruent with "dest" in *get_value*.