docs: update docs for BMI v2 (#56)
* Remove .gitignore

It's not helping anything in this minimal repository.

* Remove examples from repository

The examples will instead be linked from the README and the docs.
This may less confusing than submodules. This fixes #46.

cc: @kbarnhart

* Remove language bindings from repository

The bindings will instead be linked from the README and the docs.
This may be less confusing than submodules.

* Remove trailing .gitmodules file

* Update text; add table and links

* Strike table title

* Revise and edit text

* Add Julian Hofer as a contributor

* Update version, date, and copyright holder

I booted @mcflugen. We can always revert this.

* Update opening paragraph

* Include descriptions of BMI functions in one document

Previously, they had been linked. This will make it easier to Ctrl-F
and search for information.

* Start a glossary of terms

I'll add to it as I move through and revise the docs.

* Reformat time functions section

Still need to add function descriptions and fix duplicate header
warnings.

* A few areas where I don't understand and some typos/reformatting

* Update the model control functions section

* Include Mike Galloy as contributor

* Break update and update_until into separate sections

My plan is to list each BMI function in its own section.

* Revise and format model information functions section

* Exclude source files that are included

This removes the duplicate label warning generated when source files
with labels are included in other source files.

* Update variable info functions section

This includes moving UDUNITS links from this section and the time
functions section up to the parent document, `bmi.spec.rst`.

* Update time functions section

* Fix awkward sentences, formatting

* Create summary table of BMI functions

* Start revising descriptions of get and set functions

* Stub out section on best proactices for implementing a BMI

This is really rough right now. Also, I'm not sure where to place
it in the docs.

* List the functional categories of BMI functions

This requires adding a reference link to the heading of each category
page.

* It's 2020

* Inlcude spec and examples links at beginning; move comments to end

* Add SIDL comment to each function prototype

I don't people to confuse the function prototype with another language.

* Complete getter/setter section

* Remove the BMI Bindings section

Instead, link to the most current bindings in their respective
GitHub repos. This is one way of preventing the docs from falling
out of sync with the language specifications.

* Experiment with section navigation links

Since we don't have a floating TOC, it's difficult to move quickly
within the single-page document. To address this, I've added links
after each BMI function back to the functional category header and to
the top header. If this looks good, I'll apply it to the other
sections.

* Add section header label

* Follow file naming convention for BMI docs

I also moved this "implementation" section to Additional Topics,
and retitled it "BMI best practices".

* Checkpoint the write-up of the BMI grid functions

I'm about halfway through the BMI grid functions. I added glossary
terms, references, and a new section to describe the different grid
types supported by BMI. All these things need more work before the
discussion of model grids is complete.

* Revise section on BMI grid functions

* Minor formatting changes

* Add navigation links after each BMI function

This should make it easier to browse sections and return to the
main BMI heading.

* Display BMI language specs and examples in a table

The table replaces a cumbersome pair of bulleted lists.

* Complete revision of model grids section

* Update links and document title

* Replace examples link with link to CSDMS homepage

* Complete BMI best practices section

* Remove the beta modifier from the version

Co-authored-by: Michael Galloy <mgalloy@gmail.com>